管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -30,6 +30,8 @@ public class EmployeeAction extends ActionBase {
 
     public void index() throws ServletException, IOException {
 
+        if (checkAdmin()) {
+
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
 
@@ -47,20 +49,23 @@ public class EmployeeAction extends ActionBase {
         }
 
         forward(ForwardConst.FW_EMP_INDEX);
+        }
 
     }
     public void entryNew() throws ServletException, IOException {
+
+        if (checkAdmin()) {
 
         putRequestScope(AttributeConst.TOKEN, getTokenId());
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
         forward(ForwardConst.FW_EMP_NEW);
     }
-
+    }
     public void create() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkToken()) && checkToken()) {
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -104,6 +109,8 @@ public class EmployeeAction extends ActionBase {
 
     public void show() throws ServletException, IOException {
 
+        if (checkAdmin()) {
+
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -119,8 +126,11 @@ public class EmployeeAction extends ActionBase {
         //詳細画面を表示
         forward(ForwardConst.FW_EMP_SHOW);
     }
+    }
 
     public void edit() throws ServletException, IOException {
+
+        if (checkAdmin()) {
 
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -137,13 +147,13 @@ public class EmployeeAction extends ActionBase {
 
         //編集画面を表示する
         forward(ForwardConst.FW_EMP_EDIT);
-
+        }
     }
 
     public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkToken() && checkToken()) {
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -185,7 +195,7 @@ public class EmployeeAction extends ActionBase {
     public void destroy() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkToken() && checkToken()) {
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -196,5 +206,23 @@ public class EmployeeAction extends ActionBase {
             //一覧画面にリダイレクト
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
+    }
+
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+
+            return true;
+        }
+
     }
 }
